splitting up the mega file
