+++ src/powerquery-parser/parser/parsers/combinatorialParserV2/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export { CombinatorialParserV2 } from "./combinatorialParserV2";
