+++ src/powerquery-parser/parser/parsers/index.ts
@@ -3,6 +3,7 @@
 
 import * as NaiveParseSteps from "./naiveParseSteps";
 
+export * from "./combinatorialParserV2";
 export { NaiveParseSteps };
 export { CombinatorialParser } from "./combinatorialParserV1";
 export { RecursiveDescentParser } from "./recursiveDescentParser";
